iyp-service

+++ micro-yt-predict/src/main/java/com/yt/controller/DataController.java
@@ -23,7 +23,7 @@ public class DataController {
      * @return
      */
     @GetMapping("/getCount/{code}")
-    public List<Long> getCount(@PathVariable String code){
+    public List<Integer> getCount(@PathVariable String code){
         return ytStatistic.getCount(code);
     }
