+++ micro-yt-predict/src/main/java/com/yt/schedule/Task.java
@@ -45,13 +45,13 @@ public class Task {
     public void ask(){
         Thread thread1 = new Thread(() -> {
            while (true){
-               List<Long> count = ytStatistic.getCount("100");
+               List<Integer> count = ytStatistic.getCount("100");
                System.out.println(Thread.currentThread() + count.toString());
            }
         });
         Thread thread2 = new Thread(() -> {
             while (true){
-                List<Long> count = ytStatistic.getCount("100");
+                List<Integer> count = ytStatistic.getCount("100");
                 System.out.println(Thread.currentThread() + count.toString());
             }
         });
